requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
lightgbm
prophet
statsmodels
mlflow
plotly
streamlit
pyyaml
sqlalchemy
python-multipart
jinja2
python-dotenv
pytest
alembic
psycopg2-binary
setuptools
joblib